tests/wpt/tests/secure-payment-confirmation/authentication-accepted.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test for the 'secure-payment-confirmation' payment method authentication - accepted case</title>
<link rel="help" href="https://w3c.github.io/secure-payment-confirmation#sctn-authentication">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="utils.sub.js"></script>
<script>
'use strict';

const kChallenge = 'server challenge';
const kPayeeOrigin = 'https://merchant.com';
const kInstrumentDisplayName = 'Troycard';

async function triggerSpc(t, instrumentDetails, paymentEntitiesLogos) {
  const authenticator = await window.test_driver.add_virtual_authenticator(
      AUTHENTICATOR_OPTS);
  t.add_cleanup(() => {
    return window.test_driver.remove_virtual_authenticator(authenticator);
  });

  await window.test_driver.set_spc_transaction_mode("autoAccept");
  t.add_cleanup(() => {
    return window.test_driver.set_spc_transaction_mode("none");
  });

  const credential = await createCredential();

  let data = {
    credentialIds: [credential.rawId],
    challenge: Uint8Array.from(kChallenge, c => c.charCodeAt(0)),
    payeeOrigin: kPayeeOrigin,
    rpId: window.location.hostname,
    timeout: 60000,
    instrument: {
      displayName: kInstrumentDisplayName,
      icon: ICON_URL,
    },
  };

  if (instrumentDetails !== undefined) {
    data.instrument.details = instrumentDetails;
  }

  if (paymentEntitiesLogos !== undefined) {
    data.paymentEntitiesLogos = paymentEntitiesLogos;
  }

  const request = new PaymentRequest([{
    supportedMethods: 'secure-payment-confirmation', data
  }], PAYMENT_DETAILS);

  await test_driver.bless('user activation');
  const responsePromise = request.show();

  const response = await responsePromise;
  await response.complete('success');

  const cred = response.details;
  assert_equals(cred.id, credential.id);

  return JSON.parse(arrayBufferToString(cred.response.clientDataJSON));
}

promise_test(async t => {
  const clientDataJSON = await triggerSpc(t);

  assert_equals(clientDataJSON.type, 'payment.get');
  assert_equals(clientDataJSON.challenge, base64UrlEncode(kChallenge));
  assert_equals(clientDataJSON.origin, window.location.origin);
  assert_false(clientDataJSON.crossOrigin);

  // Payment-specific information.
  assert_equals(clientDataJSON.payment.rpId, window.location.hostname);
  assert_equals(clientDataJSON.payment.topOrigin, window.location.origin);
  assert_equals(clientDataJSON.payment.payeeOrigin, kPayeeOrigin);
  assert_equals(clientDataJSON.payment.total.value, PAYMENT_DETAILS.total.amount.value);
  assert_equals(clientDataJSON.payment.total.currency, PAYMENT_DETAILS.total.amount.currency);
  assert_equals(clientDataJSON.payment.instrument.icon, ICON_URL);
  assert_equals(clientDataJSON.payment.instrument.displayName, kInstrumentDisplayName);

  // If the User Agent still supports the legacy 'rp' output parameter, it
  // should be identical to the 'rpId' output parameter. See
  // https://github.com/w3c/secure-payment-confirmation/pull/198
  if ('rp' in clientDataJSON.payment) {
    assert_equals(clientDataJSON.payment.rp, clientDataJSON.payment.rpId);
  }

  // TODO: Verify cred.response.signature, to validate that it covers all fields
  // from clientDataJSON.
}, 'Successful SPC authentication - mandatory fields');

promise_test(async t => {
  const instrumentDetails = '***1234';
  const paymentEntityLogoLabel = 'Sync Network';
  const paymentEntitiesLogos = [{
      url: PAYMENT_ENTITY_LOGO_URL,
      label: paymentEntityLogoLabel,
  }];
  const clientDataJSON = await triggerSpc(t, instrumentDetails, paymentEntitiesLogos);

  // Payment-specific information.
  assert_equals(clientDataJSON.payment.instrument.details, instrumentDetails);
  assert_equals(clientDataJSON.payment.paymentEntitiesLogos.length, 1);
  assert_equals(clientDataJSON.payment.paymentEntitiesLogos[0].url, PAYMENT_ENTITY_LOGO_URL);
  assert_equals(clientDataJSON.payment.paymentEntitiesLogos[0].label, paymentEntityLogoLabel);
}, 'Successful SPC authentication - optional fields');
</script>
